check if it is live mode before cleaning up autotalent

+++ src/com/intervigil/micdroid/Mic.java
@@ -326,7 +326,9 @@ public class Mic extends Activity implements OnClickListener {
     DependentTask postRecordTask = new DependentTask() {
         @Override
         public void doTask() {
-            Autotalent.destroyAutotalent();
+            if (PreferenceHelper.getLiveMode(Mic.this)) {
+                Autotalent.destroyAutotalent();
+            }
             Toast.makeText(getBaseContext(),
                     R.string.recording_finished_toast,
                     Toast.LENGTH_SHORT).show();
@@ -338,7 +340,9 @@ public class Mic extends Activity implements OnClickListener {
 
         @Override
         public void handleError() {
-            Autotalent.destroyAutotalent();
+            if (PreferenceHelper.getLiveMode(Mic.this)) {
+                Autotalent.destroyAutotalent();
+            }
             recordingButton.setOnCheckedChangeListener(null);
             recordingButton.setChecked(false);
             recordingButton.setOnCheckedChangeListener(recordBtnListener);
